Qlearning_to_realize_cart_pole_balance_game
Q learning principle and Implementation (CART pole balance game)

There is an environment described by state s (s belongs to s, s is the set of all possible states), an agent that can execute action a,

(a belongs to a, and a is the set of all possible actions). The action of the agent causes the agent to transfer from one state to another. The behavior of the agent will be rewarded, and the behavior of the agent

The goal is to maximize rewards.

In Q learning, the agent calculates the state action combination that can maximize the reward R, so as to learn the action to be taken (strategy PI). When selecting the action,

The agent should not only consider the current reward, but also try to consider the future reward